pyproject.toml
[project]
name = "pdf2zh"
version = "1.9.6.post13"
description = "Latex PDF Translator"
authors = [{ name = "Byaidu", email = "byaidux@gmail.com" }]
license = "AGPL-3.0"
readme = "README.md"
requires-python = ">=3.10,<3.13"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
dependencies = [
    "requests",
    "pymupdf>=1.24",
    "tqdm",
    "tenacity",
    "numpy",
    "openai",
    "huggingface_hub",
    "onnx",
    "onnxruntime",
    "opencv-python-headless",
    # https://github.com/pdfminer/pdfminer.six/issues/1061
    # https://github.com/pdfminer/pdfminer.six/issues/1110
    "pdfminer.six==20240706",
    "peewee>=3.17.8",
    "fontTools",
    "babeldoc>=0.1.22, <0.3.0",
    "rich",
    "tiktoken>=0.9.0",
]

[dependency-groups]
dev = [
    "ruff==0.11.0",
    "pre-commit",
    "pytest",
    "build",
    "lxml-stubs>=0.5.1",
    "pandas-stubs>=2.2.3.250308",
    "types-pyyaml>=6.0.12.20241230",
    "types-pygments>=2.19.0.20250305",
    "types-aiofiles>=24.1.0.20241221",
    "types-colorama>=0.4.15.20240311",
    "types-defusedxml>=0.7.0.20240218",
    "types-peewee>=3.17.9.20250308",
    "types-protobuf>=5.29.1.20250315",
    "types-pycurl>=7.45.6.20250309",
    "types-pytz>=2025.1.0.20250318",
    "types-requests>=2.32.0.20250306",
    "types-simplejson>=3.20.0.20250318",
    "types-tensorflow>=2.18.0.20250319",
    "types-tqdm>=4.67.0.20250319",
    "types-ujson>=5.10.0.20240515",
    "bumpver>=2024.1130",
]

[project.optional-dependencies]
compatible = ["pikepdf"]
gui = ["gradio", "gradio_pdf>=0.0.21"]
argostranslate = ["argostranslate"]
azure-ai = ["azure-ai-translation-text<=1.0.1"]
deepl = ["deepl"]
ollama = ["ollama"]
tencent = ["tencentcloud-sdk-python-tmt"]
xinference = ["xinference-client"]

[project.urls]
Homepage = "https://github.com/Byaidu/PDFMathTranslate"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
pdf2zh = "pdf2zh.pdf2zh:main"

[tool.flake8]
ignore = ["E203", "E261", "E501", "W503", "E741"]
max-line-length = 88

[tool.uv]
managed = true
package = true
no-sources = false
override-dependencies = ["pymupdf>=1.24"]

[tool.ruff]
line-length = 120
indent-width = 4
target-version = "py310"

[tool.ruff.lint]
select = [
    # pycodestyle
    "E",
    # Pyflakes
    "F",
    # flake8-bugbear
    "B",
    "W",
    # flake8-future-annotations
    "FA",
    # isort
    "I",
    # pyupgrade
    "UP",
]
ignore = [
    "B006",
    "E203",
    "E261",
    "E501",
    "E741",
    "D100",
    "D203",
    "D213",
    "D300",
    "E111",
    "E114",
    "E117",
    "ISC001",
    "ISC002",
    "W191",
]

[tool.ruff.lint.isort]
extra-standard-library = ["typing_extensions"]


[bumpver]
current_version = "1.9.6.post13"
version_pattern = "MAJOR.MINOR.PATCH[.PYTAGNUM]"

[bumpver.file_patterns]
"pyproject.toml" = ['current_version = "{version}"', 'version = "{version}"']
"pdf2zh/__init__.py" = ['__version__ = "{version}"']
